move more options
