+++ tools/build_variables.py
@@ -197,8 +197,12 @@ def add_torch_libs():
         "torch/csrc/api/src/nn/modules/rnn.cpp",
         "torch/csrc/api/src/nn/modules/container/functional.cpp",
         "torch/csrc/api/src/nn/modules/container/named_any.cpp",
+        "torch/csrc/api/src/nn/options/batchnorm.cpp",
         "torch/csrc/api/src/nn/options/conv.cpp",
+        "torch/csrc/api/src/nn/options/dropout.cpp",
+        "torch/csrc/api/src/nn/options/linear.cpp",
         "torch/csrc/api/src/nn/options/pooling.cpp",
+        "torch/csrc/api/src/nn/options/rnn.cpp",
         "torch/csrc/api/src/optim/adagrad.cpp",
         "torch/csrc/api/src/optim/adam.cpp",
         "torch/csrc/api/src/optim/lbfgs.cpp",
