+++ torch/csrc/api/include/torch/nn/options.h
@@ -1,3 +1,10 @@
 #pragma once
 
+#include <torch/nn/options/batchnorm.h>
+#include <torch/nn/options/conv.h>
+#include <torch/nn/options/dropout.h>
+#include <torch/nn/options/fold.h>
+#include <torch/nn/options/linear.h>
+#include <torch/nn/options/loss.h>
 #include <torch/nn/options/pooling.h>
+#include <torch/nn/options/rnn.h>
