+++ torch/csrc/api/include/torch/nn/options/conv.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include <torch/arg.h>
+#include <torch/csrc/WindowsTorchApiMacro.h>
 #include <torch/expanding_array.h>
+#include <torch/types.h>
 
 namespace torch {
 namespace nn {
