+++ torch/csrc/api/include/torch/nn/options/loss.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <torch/arg.h>
+#include <torch/csrc/WindowsTorchApiMacro.h>
+#include <torch/types.h>
 
 namespace torch {
 namespace nn {
