+++ torch/csrc/api/include/torch/nn/options/pooling.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <torch/arg.h>
+#include <torch/csrc/WindowsTorchApiMacro.h>
 #include <torch/expanding_array.h>
 #include <torch/types.h>
