+++ torch/csrc/api/src/nn/options/dropout.cpp
@@ -1,0 +1,9 @@
+#include <torch/nn/options/dropout.h>
+
+namespace torch {
+namespace nn {
+
+DropoutOptions::DropoutOptions(double rate) : rate_(rate) {}
+
+} // namespace nn
+} // namespace torch
