+++ torch/csrc/api/src/nn/options/pooling.cpp
@@ -3,6 +3,10 @@
 namespace torch {
 namespace nn {
 
+template struct AvgPoolOptions<1>;
+template struct AvgPoolOptions<2>;
+template struct AvgPoolOptions<3>;
+
 template struct MaxPoolOptions<1>;
 template struct MaxPoolOptions<2>;
 template struct MaxPoolOptions<3>;
